더미데이터 리펙토링
